outlineLevel type fix

+++ src/Gooseberry.ExcelStreaming/Attributes/RowAttributes.cs
@@ -5,6 +5,6 @@ namespace Gooseberry.ExcelStreaming.Attributes;
 [StructLayout(LayoutKind.Auto)]
 public readonly record struct RowAttributes(
     decimal? Height = null,
-    uint? OutlineLevel = null,
+    byte? OutlineLevel = null,
     bool IsHidden = false,
     bool IsCollapsed = false);
